Disentangle false inheritance relationship
A RelatedResource is only a Description, it does not have the other Cocina schemas
